SQUARE PEG THROUGH ROUND HOLE ACHIEVED
Was too worried about enforcing structure with last rendition. Literally put in code to prevent creation of cyclic bonds. Did a quick rewrite, we are good now. I did a half as verbose as it was and half how I would prefer to throw it together. Going to read over paper now and see if it alters my possible approaches.