added graph structure
Thinking, probably going to have our graph class just contain our list of nodes present in this specific graph.  This may be nice to template for peoples brains (i.e. molecule class) but it does add more templating which is ass. It all works proper, bug tested last night. Next step will be to allow our edges to be hashed to specific graphs, this way our nodes are shared between structures but the connectivity/edges are not. 

Most likely will just have an overload for current functionality that takes graph then somehow use this as a key across functions. Have a couple ideas.