Added beginnings of needed functions
Added our declares of functions that will be needed to allow nodes to be shared among different structures.  

Current implementation: Our graph is given a a unique id (just going to use our memory address casted to std::uintptr_t (variable size uint that is automatically large enough to store our mem address as a prim type) which will be stored at graph creation) which is hashed for the key value. The largest current worry is do we need our nodes & edges to know which graph they are a part of? I am assuming NO as of now but this is subject to change. 

Note: As of now, the hardest part is taken care of, mostly regarding the logic and proper implementation of adding, deleting, etc. nodes. Further changes will be less intensive because they will most likely be changing relations (i.e. std::shared_ptr to std::weak_ptr) or placing data into containers (i.e. for our specific edges being placed into a map that contains the in & out edges, please note that we may end up creating a custom struct or tuple to make this a little cleaner, currently the separation/not using a tuple is used on purpose in order to make it easier to understand what is going on since there is so much abstraction as is)